commit code 12/4

+++ AnService_Capstone.Core/Entities/TblUsedMaterial.cs
@@ -11,8 +11,8 @@ namespace AnService_Capstone.Core.Entities
         public int? MaterialId { get; set; }
         public int? RequestDetailId { get; set; }
         public int? WorkerId { get; set; }
-        public int? Quantity { get; set; }
-        public int? QuantityNew { get; set; }
+        public double? Quantity { get; set; }
+        public double? QuantityNew { get; set; }
         public int? Status { get; set; }
         public string Note { get; set; }
         public string Message { get; set; }
